Add property deprecation
